refactor: remove perf files
