+++ crates/risc0/src/lib.rs
@@ -14,6 +14,3 @@ pub use groth16_verifier::*;
 
 mod env;
 pub use env::Risc0ZkVmEnv;
-
-#[cfg(feature = "perf")]
-mod perf;
