+++ crates/sp1/src/lib.rs
@@ -15,8 +15,5 @@ mod env;
 #[cfg(feature = "zkvm")]
 pub use env::Sp1ZkVmEnv;
 
-#[cfg(feature = "perf")]
-pub mod perf;
-
 mod groth16_verifier;
 pub use groth16_verifier::*;
