+++ crates/zkvm/src/prover.rs
@@ -1,5 +1,7 @@
+use std::time::Instant;
+
 use crate::{
-    host::ZkVmHost, input::ZkVmInputBuilder, ProofReceipt, ProofType, PublicValues,
+    host::ZkVmHost, input::ZkVmInputBuilder, ProofReceipt, ProofReport, ProofType, PublicValues,
     ZkVmInputResult, ZkVmResult,
 };
 
@@ -36,6 +38,7 @@ pub trait ZkVmProver {
         // Process output to see if we are getting the expected type.
         let _ = Self::process_output::<H>(receipt.public_values())?;
 
+        // Dump the proof to file if flag is enabled
         if std::env::var("ZKVM_PROOF_DUMP")
             .map(|v| v == "1" || v.to_lowercase() == "true")
             .unwrap_or(false)
@@ -45,5 +48,39 @@ pub trait ZkVmProver {
         }
 
         Ok(receipt)
+    }
+
+    /// Generates the proof report using any zkVM host.
+    fn perf_report<'a, H>(input: &'a Self::Input, host: &H) -> ZkVmResult<ProofReport>
+    where
+        H: ZkVmHost,
+        H::Input<'a>: ZkVmInputBuilder<'a>,
+    {
+        let start = Instant::now();
+
+        // Prepare the input using the host's input builder.
+        let zkvm_input = Self::prepare_input::<H::Input<'a>>(input)?;
+
+        let (_, cycles) = host.execute(zkvm_input)?;
+        let execution_time = start.elapsed().as_millis();
+
+        if std::env::var("ZKVM_PROFILING_DUMP")
+            .map(|v| v == "1" || v.to_lowercase() == "true")
+            .unwrap_or(false)
+        {
+            let from = format!("{:?}.profle", host);
+            let to = format!("{}_{:?}.profile", Self::name(), host);
+            std::fs::rename(from, to).unwrap();
+        }
+
+        let _ = Self::prove(input, host)?;
+        let proving_time = start.elapsed().as_millis();
+
+        Ok(ProofReport {
+            name: Self::name(),
+            cycles,
+            execution_time,
+            proving_time,
+        })
     }
 }
