+++ examples/schnorr-sig-verify/src/prover.rs
@@ -1,4 +1,4 @@
-use strata_zkvm::{ProofType, ZkVmEnv, ZkVmInputResult, ZkVmProver, ZkVmProverPerf};
+use strata_zkvm::{ProofType, ZkVmEnv, ZkVmInputResult, ZkVmProver};
 
 use crate::{verify_schnorr_sig_k256, SchnorrSigInput};
 
@@ -46,8 +46,6 @@ impl ZkVmProver for SchnorrSigProver {
         H::extract_serde_public_output(public_values)
     }
 }
-
-impl ZkVmProverPerf for SchnorrSigProver {}
 
 #[cfg(test)]
 mod tests {
