+++ runner/src/programs/sha2.rs
@@ -1,7 +1,7 @@
 use sha2_chain::ShaChainProver;
-use strata_zkvm::{ProofReport, ZkVmHostPerf, ZkVmProver, ZkVmProverPerf};
+use strata_zkvm::{ProofReport, ZkVmHost, ZkVmProver};
 
-fn sha2_prover_perf_report(host: &impl ZkVmHostPerf) -> ProofReport {
+fn sha2_prover_perf_report(host: &impl ZkVmHost) -> ProofReport {
     let input = 5;
     ShaChainProver::prove(&input, host).unwrap();
     ShaChainProver::perf_report(&input, host).unwrap()
